bot/requirements.txt
aiogram==3.12.0
aiogram_dialog==2.2.0b3
redis==5.0.8
cryptography==42.0.8
ujson==5.10.0
python-dotenv==1.0.1
tortoise-orm==0.21.5
asyncpg==0.29.0
openpyxl==3.1.2
